feat: enhance SalesforceToolingClient tests with shared auth and async apex job queries
